Revert "Fixing implicit conversion changes to signedness"
This reverts commit 7099a4437e84740d2e5cc6242561b3fd894ff802.

+++ src/logger.c
@@ -96,8 +96,8 @@ static void real_logger(int level, int priority, const char *message) {
 			logcontrol = true;
 			if(level > (c->outcompression >= 0 ? (int)c->outcompression : (int)debug_level))
 				continue;
-			size_t len = strlen(message);
-			if(send_request(c, "%d %d %zu", CONTROL, REQ_LOG, len))
+			int len = strlen(message);
+			if(send_request(c, "%d %d %d", CONTROL, REQ_LOG, len))
 				send_meta(c, message, len);
 		}
 		suppress = false;
